backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
rdflib>=7.0.0
groq>=0.4.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
chromadb>=0.4.18
requests>=2.31.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0